[C-Warning] implicit declaration of function 'rb_tinytds_raise_error'
When I fixed this by moving things around, I these at one point or another.

  incompatible integer to pointer conversion passing 'char' to parameter of type 'const void *' [-Wint-conversion]
    strcpy(userdata->nonblocking_error.error, dberrstr);

  incompatible pointer types passing 'char (*)[1024]' to parameter of type 'char *' [-Wincompatible-pointer-types]
    &userdata->nonblocking_error.error,

  incompatible pointer types passing 'char **' to parameter of type 'char *'; remove & [-Wincompatible-pointer-types]
    &userdata->nonblocking_error.error,

No I get random segmentation faults but I can not tell if it is due to this or the now exposed warnings for `NOGVL_DBCALL` functions.
